`main` moved thirteen commits under this branch, from `c3522bc7d` to `25ee19464`,
and the last of them is W5d-1 (#2249 item 1, #2265), which edits three of the
four files this wave also edits. This is a product merge, not a records refresh.

NOTHING IN THIS COMMIT HAS BEEN BUILT OR RUN. The box is at 204 MB free with the
filesystem at 100%, other sessions are compiling, and a build begun at 872 MB was
killed part-way. AGENTS.md and this row's own record both refuse a green computed
on a full disk: under ENOSPC a failed build leaves the previous binary in place
and reads as a PASSING test, and this tree's checkers emit FALSE policy refusals.
So the resolution below is stated as a resolution and NOT as a verified one. It
is committed to preserve the reconciliation, not to claim it passes.
`scripts/agent-preflight.sh` has NOT been run at this head, no suite has been
run, and this branch is NOT pushed.

WHAT THE MERGE ACTUALLY DECIDED, AND WHY GIT COULD NOT DECIDE IT.
`src/vllm/model_executor/models/qwen4_exp_registry.cpp` conflicted, and the
conflict is semantic rather than textual. The production refusal enumerates the
prerequisites the layer loop still lacks. The survey listed FIVE. #2265 removed
items (1), the grouped RMS norm, and (5), the mRoPE builder, and renumbered five
to THREE — while still listing the paged QSA consumer, which THIS wave closes.
W5d-3 folded item (2) into item (3) and renumbered five to FOUR — while still
listing the grouped RMS norm and the mRoPE builder, which #2265 closes. Each
edit is correct against the baseline it was written on and WRONG at this merge.
Taking either side whole lands a refusal that names finished work, which is
#2288 for the third time on this row in one day, and no gate can catch it: the
`SUBCASE("the forward")` of `test_qwen4_exp_scaffold.cpp:767` pins that five
substrings are PRESENT, never that the enumeration is TRUE.

The resolution is therefore the SET DIFFERENCE of the survey against BOTH waves,
not the shorter of the two lists. THREE of the five are closed — item 1 by W5d-1,
item 5 by W5d-2, item 2 by W5d-3 for its K/V half — and TWO remain: reach for the
indexer side cache, whose group-2 block table `gather_block_table` never gathers
(W5c-2), and the MoE weight adapter. The landed message names exactly those two,
carries the merged wave list `W5b-1..6, W5c-1, W5d-1, W5d-2, W5d-3`, and keeps
W5d-3's wording for the surviving item because it is the wording that records
which half was discharged. All five substrings the scaffold suite pins are still
in the string, and "was not produced by" still appears only in a comment and
never in the message, so the asserted absence holds. THAT IS A READING OF THE
BYTES, NOT A TEST RESULT, because no test was run.

`.agents/specs/qwen4-exp-flash-next.md` conflicted and was resolved as a SET
against the merge base `c3522bc7d`, per AGENTS.md `## Records`, which refuses an
automatic three-way merge of a keyed record. Both sides added at the SAME anchor
and both are kept: this branch's `## Mutation record — W5d-3` follows `main`'s
W5b-6, W5d-2 and W5d-1 sections in landing order. Three places collided in
substance and each now carries BOTH facts — the `## Now` count paragraph, which
said TWO closed on one side and FOUR remaining on the other and now says three
closed and two remaining with the arithmetic shown; the `## Owed` survey item 2,
now marked CLOSED for its K/V half with the survey argument kept because the
layer loop still has to CALL the paged arm; and the `## Port map` PAGED-cache
bullet, struck through and marked HALF DISCHARGED. Against `origin/main` the
resolved spec is five hunks with twenty deletions, and every deletion is one of
those three replacements.

`include/vt/ops.h` and `src/vt/ops.cpp` merged without a conflict and were
checked rather than trusted: every symbol #2265 landed is present in the merged
tree — `kRmsNormGroup`, `RmsNormGroupArgs`, `RmsNormGroupFn`, `vt::RmsNormGroup`,
the dispatcher, the CPU kernel, the `op_provider` name entry, the
`test_ops_rms_norm_group` target and its source file. The `OpId` enum has no
duplicate enumerator and `kRmsNormGroup` is still the last before `kCount`, so
its id has not shifted. `.agents/issue-index.md` did not conflict, this branch
appends no row, and the merged file is byte-identical to `origin/main`: 868 rows,
zero duplicate row ids matched on `^| \[#NNNN\]`, exactly one `#2249` row.

The message is written rather than taken from `git merge --no-edit`, because
`check-commit-trailers.py` walks `rev-list` without `--no-merges` and a
trailer-less merge commit has no non-force repair.

FOLLOWING_AGENTS_PROTOCOL

Following-Agents-Protocol: true
AI-Assisted: true

+++ include/vllm/model_executor/models/deepseek_v4_device.h
@@ -95,9 +95,14 @@ struct DsaDeviceKernels {
                                const std::vector<int64_t>& win_end, int64_t num_tokens,
                                int64_t num_keys, int64_t topk);
   std::vector<float> (*softmax_sink)(vt::Queue&, const std::vector<float>& scores, float sink);
+  // W1d (#2186): the two weights are BF16 bit patterns (`HostBf16`), not f32 --
+  // the carried tower's FP8-sourced half is held at the model dtype. A function
+  // pointer cannot be a template, so this entry names the ONE dtype the carried
+  // tower actually has; the f32 arm lives on in the CPU `GroupedOutputLora<float>`
+  // that the ported upstream-parity tests drive.
   std::vector<float> (*grouped_olora)(vt::Queue&, const std::vector<float>& o,
-                                      const std::vector<float>& wo_a,
-                                      const std::vector<float>& wo_b, int64_t num_tokens,
+                                      const std::vector<uint16_t>& wo_a,
+                                      const std::vector<uint16_t>& wo_b, int64_t num_tokens,
                                       int64_t n_heads, int64_t head_dim, int64_t n_groups,
                                       int64_t o_lora_rank, int64_t hidden_size);
   // Brick A — device MLA decode/prefill attention over the unified KV-cache latent.
